requirements.txt
Flask==2.1.1
torch==1.11.0
torchvision==0.12.0
Pillow==9.1.1
Werkzeug==2.0.3